+++ src/main/java/com/shy_polarbear/server/domain/user/controller/AuthController.java
@@ -1,6 +1,6 @@
 package com.shy_polarbear.server.domain.user.controller;
 
-import com.shy_polarbear.server.domain.config.jwt.JwtDto;
+import com.shy_polarbear.server.global.config.jwt.JwtDto;
 import com.shy_polarbear.server.domain.user.dto.*;
 import com.shy_polarbear.server.domain.user.dto.JoinRequest;
 import com.shy_polarbear.server.domain.user.dto.SocialLoginRequest;
