+++ src/main/java/com/shy_polarbear/server/global/config/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.config.jwt;
+package com.shy_polarbear.server.global.config.jwt;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
