+++ src/main/java/com/shy_polarbear/server/global/config/security/SecurityConfig.java
@@ -1,16 +1,15 @@
-package com.shy_polarbear.server.domain.config.security;
+package com.shy_polarbear.server.global.config.security;
 
 
 
-import com.shy_polarbear.server.domain.config.jwt.JwtAuthenticationFilter;
-import com.shy_polarbear.server.domain.config.jwt.JwtProvider;
+import com.shy_polarbear.server.global.config.jwt.JwtAuthenticationFilter;
+import com.shy_polarbear.server.global.config.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
